+++ include/nuttx/clk/clk.h
@@ -75,16 +75,16 @@ int      clk_enable(struct clk *clk);
 void     clk_disable(struct clk *clk);
 int      clk_is_enabled(struct clk *clk);
 
-int64_t  clk_round_rate(struct clk *clk, uint64_t rate);
-int      clk_set_rate(struct clk *clk, uint64_t rate);
-uint64_t clk_get_rate(struct clk *clk);
+uint32_t clk_round_rate(struct clk *clk, uint32_t rate);
+int      clk_set_rate(struct clk *clk, uint32_t rate);
+uint32_t clk_get_rate(struct clk *clk);
 
 int      clk_set_phase(struct clk *clk, int degrees);
 int      clk_get_phase(struct clk *clk);
 
 /* Clk helper API */
 
-const char* clk_get_name(const struct clk *clk);
+const char *clk_get_name(const struct clk *clk);
 
 #undef EXTERN
 #ifdef __cplusplus
